* master: (38 commits)
  Update Go Bindings
  CLEANFILES += *.class
  Fixup language tests.
  Detect which ruby dir to use
  Update test env with RUBYLIB
  Extend timing for test runner.
  Fix admin perf counters on OS X
  Improve compatibility with newer automake
  Fix hyperleveldb version
  Solve type mismatch on 64-bit Darwin and gcc 4.9.
  Make distcheck for Python, Java
  Tests for non-equality search.
  Fix ruby special predicates
  Fix the secondary indices on integers.
  Prefer any index over a full scan, unless costs show it is unwise
  Output description of search with search_describe
  Add basic search test.
  Fix ruby maps
  Fix python maps of floats
  Fix float/int pack/unpack routines.
  ...

Conflicts:
	configure.ac

+++ daemon/identifier_collector.cc
@@ -175,7 +175,7 @@ identifier_collector :: adopt(region_id* ris, size_t ris_sz)
 
     // get the old counters
     counter* old_lower_bounds = NULL;
-    size_t old_sz = 0;
+    uint64_t old_sz = 0;
     get_base(&old_lower_bounds, &old_sz);
 
     // carry over values
